use graphical object instead of text in OrderedListView
